# docs/api/procurement.openapi.yaml
openapi: 3.1.0
info:
  title: Construction OS — Procurement Service API
  version: 1.0.0
  description: 'Phase 5 Procurement Service. Covers vendor management, purchase requests,

    RFQ lifecycle, quotation comparison, PO approval chain (Temporal-backed),

    delivery recording, and invoice receipt.


    Approval thresholds (spec §15.5):

    - ≤ 50,000 THB: PROJECT_MANAGER alone

    - 50,001–500,000 THB: PROJECT_MANAGER + FINANCE

    - > 500,000 THB: PROJECT_MANAGER + FINANCE + EXECUTIVE

    - 48h timeout per approver → escalates to TENANT_ADMIN

    '
servers:
  - url: /api/v1
    description: Current API version
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CreateVendorDto:
      type: object
      required:
        - vendor_code
        - vendor_name
      properties:
        vendor_code:
          type: string
          maxLength: 50
        vendor_name:
          type: string
          maxLength: 255
        tax_id:
          type: string
          maxLength: 100
          nullable: true
        contact_email:
          type: string
          format: email
        contact_phone:
          type: string
          maxLength: 50
        address:
          type: string
...
    CreateRfqDto:
      type: object
      required:
        - project_id
        - rfq_number
        - deadline
      properties:
        project_id:
          type: string
          format: uuid
        pr_id:
          type: string
          format: uuid
        rfq_number:
          type: string
          maxLength: 50
        deadline:
          type: string
          format: date-time
...
paths:
  /procurement/vendors:
    post:
      tags:
        - Vendors
      summary: Create a new vendor
      operationId: createVendor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateVendorDto'
      responses:
        '201':
          description: Vendor created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VendorRow'
        '422':
          description: Business rule violation
    get:
      tags:
        - Vendors
      summary: List vendors
      operationId: listVendors
      parameters:
        - name: active_only
          in: query
          required: false
          schema:
            type: boolean
            default: true
      responses:
        '200':
          description: Vendor list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/VendorRow'
  /procurement/vendors/directory:
    get:
      tags:
        - Vendors
      summary: Vendor directory — active vendors with their open-project count
      description: >-
        Serves the mobile vendor directory
        (drawn in the `mockup/mobile/06_project_manager` set withdrawn on 2026-08-10; the vendor
        directory now hangs off the More menu). Each row is the vendor
        plus `active_project_count`. Separate from GET /procurement/vendors, which is the plain
        vendor master read and does not pay for the per-row aggregate. The trust score is NOT
        included: it is a per-vendor computation, GET /procurement/vendors/{vendorId}/score.
      operationId: listVendorDirectory
      parameters:
        - name: category
          in: query
          required: false
          description: >-
            Narrow to one category. An unknown value returns 400 — a filter that silently ignores
            its own argument would return a full list the caller believes is filtered.
          schema:
            type: string
            enum: [MATERIALS, LOGISTICS, SERVICES, EQUIPMENT]
      responses:
        '200':
          description: Active vendors, alphabetical by name
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/VendorDirectoryRow'
        '400':
          description: Unknown category
  /procurement/vendors/{vendorId}:
    get:
      tags:
        - Vendors
      summary: Get vendor by ID
      operationId: getVendor
      parameters:
        - name: vendorId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VendorRow'
        '404':
          description: Not found
    delete:
      tags:
        - Vendors
      summary: Deactivate vendor (soft delete)
      operationId: deactivateVendor
      parameters:
        - name: vendorId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Deactivated
        '404':
          description: Not found
  /procurement/vendors/{vendorId}/quotations:
    get:
      tags:
        - Vendors
      summary: List a vendor's quotation history (all RFQs, newest first)
      operationId: getVendorQuotations
      parameters:
        - name: vendorId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: The vendor's quotations across all RFQs, ordered by submitted_at DESC
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/QuotationRow'
        '404':
          description: Vendor not found
  /procurement/rfqs:
    get:
      tags:
        - RFQs
      summary: List RFQs across the tenant (filterable) — AIP-132
      operationId: listAllRfqs
      parameters:
        - name: project_id
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/RfqRow'
                  total:
                    type: integer
                  page:
                    type: integer
                  limit:
                    type: integer
    post:
      tags:
        - RFQs
      summary: Create RFQ and start Temporal workflow
      operationId: createRfq
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRfqDto'
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RfqRow'
  /procurement/rfqs/{rfqId}/publish:
    post:
      tags:
        - RFQs
      summary: Publish RFQ (DRAFT → PUBLISHED)
      operationId: publishRfq
      parameters:
        - name: rfqId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Signal sent
        '422':
          description: Not in DRAFT state
  /procurement/rfqs/{rfqId}/close:
    post:
      tags:
        - RFQs
      summary: Manually close RFQ before deadline (PUBLISHED → CLOSED)
      operationId: closeRfq
      parameters:
        - name: rfqId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Signal sent
  /procurement/rfqs/{rfqId}/cancel:
    post:
      tags:
        - RFQs
      summary: Cancel RFQ
      operationId: cancelRfq
      parameters:
        - name: rfqId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Signal sent
  /procurement/rfqs/{rfqId}/award:
    post:
      tags:
        - RFQs
      summary: Award RFQ to selected quotation (EVALUATED → AWARDED)
      operationId: awardRfq
      parameters:
        - name: rfqId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - quotation_id
              properties:
                quotation_id:
                  type: string
                  format: uuid
      responses:
        '200':
          description: RFQ awarded
        '404':
          description: Quotation not found in RFQ
  /procurement/rfqs/{rfqId}/quotations:
    post:
      tags:
        - Quotations
      summary: Submit a vendor quotation for an RFQ
      operationId: submitQuotation
      parameters:
        - name: rfqId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitQuotationDto'
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuotationRow'
        '422':
          description: RFQ not PUBLISHED
    get:
      tags:
        - Quotations
      summary: Compare quotations (sorted by price ASC, lowest auto-selected)
      operationId: compareQuotations
      parameters:
        - name: rfqId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/QuotationRow'
        '422':
          description: RFQ not CLOSED or no quotations
  /procurement/purchase-orders:
    get:
      tags:
        - Purchase Orders
      summary: List purchase orders across the tenant (filterable) — AIP-132
      operationId: listAllPurchaseOrders
      parameters:
        - name: project_id
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/PurchaseOrderRow'
                  total:
                    type: integer
                  page:
                    type: integer
                  limit:
                    type: integer
    post:
      tags:
        - Purchase Orders
      summary: Create PO and start Temporal workflow
      operationId: createPurchaseOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePurchaseOrderDto'
      responses:
        '201':
          content:
            application/json:
              schema:
                type: object
                properties:
                  po:
                    $ref: '#/components/schemas/PurchaseOrderRow'
                  line_items:
                    type: array
                    items:
                      $ref: '#/components/schemas/PoLineItemRow'
        '422':
          description: Total mismatch or validation error
  /procurement/purchase-orders/{poId}:
    get:
      tags:
        - Purchase Orders
      summary: Get PO detail with line items
      operationId: getPurchaseOrder
      parameters:
        - name: poId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  po:
                    $ref: '#/components/schemas/PurchaseOrderRow'
                  line_items:
                    type: array
                    items:
                      $ref: '#/components/schemas/PoLineItemRow'
  /procurement/purchase-orders/{poId}/submit:
    post:
      tags:
        - Purchase Orders
      summary: Submit PO for approval (DRAFT → PENDING_APPROVAL)
      operationId: submitPoForApproval
      parameters:
        - name: poId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Signal sent
  /procurement/purchase-orders/{poId}/approve:
    post:
      tags:
        - Purchase Orders
      summary: Approve PO for a tier (PM / FINANCE / EXECUTIVE / TENANT_ADMIN)
      operationId: approvePo
      parameters:
        - name: poId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - tier
              properties:
                tier:
                  type: string
                  enum:
                    - PM
                    - FINANCE
                    - EXECUTIVE
                    - TENANT_ADMIN
      responses:
        '200':
          description: Approval signal sent
...
  /procurement/purchase-orders/{poId}/acknowledge:
    post:
      tags:
        - Purchase Orders
      summary: Record vendor acknowledgement (SENT → ACKNOWLEDGED)
      operationId: acknowledgePo
      parameters:
        - name: poId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Signal sent
...